+++ src/uni_electrolyte/evaluator/inference/pyG_entry.py
@@ -7,7 +7,7 @@ import torch
 from ase.db import connect
 from torch_geometric.data import DataLoader
 from uni_electrolyte.evaluator.dataset import thuEMol
-from uni_electrolyte.evaluator.dataset import xyz_2_db, smile_2_db, db_2_pyG_data
+from uni_electrolyte.evaluator.dataset import xyz_2_db, smile_2_db, db_2_pyG_data, edit_ase_db
 from uni_electrolyte.evaluator.inference import pyG_inference_without_label
 from uni_electrolyte.evaluator.model.spatial import LEFTNet
 from uni_electrolyte.evaluator.trainer import pyG_trainer
@@ -18,6 +18,12 @@ decorated_property_names = {
     'lumo': 'LUMO(eV)',
     'dielectric_constant': 'Dielectric constant',
     'viscosity': 'Viscosity (mPa*s)',
+}
+
+three_property_names = {
+    'sp': 'Solvating Power',
+    'rs': 'Reductive Stability',
+    'sa': 'SEI-fluorinating Ability',
 }
 
 decorated_property_names_reverse = {
@@ -34,16 +40,36 @@ decorated_property_names_reverse = dict(
 
 leftnet_param = {
     'num_layers': 6,
-    'hidden_channels': 192,
+    'hidden_channels': 128,
     'num_radial': 96,
     'cutoff': 8,
 }
+
+
+def screen_csv_with_conditions(input_csv_abs_path: str, output_csv_abs_path: str, conditions_dict: dict):
+    input_df = pd.read_csv(input_csv_abs_path)
+    for column, value_range in conditions_dict.items():
+        ouput_df = input_df[(input_df[column] >= value_range[0]) & (input_df[column] <= value_range[1])]
+    ouput_df.to_csv(output_csv_abs_path, index=False)
+
 
 def pyG_infer_from_db(db_path, target_list, eval_ckpt_path, output_directory,
                       abs_data_path=None, infer_batch_size=50, use_sigmoid=False,
                       leftnet_param=leftnet_param, decorated_property_names=decorated_property_names):
     if not abs_data_path:
         abs_data_path = output_directory
+    output_directory = os.path.abspath(output_directory)
+
+    # for Early stop and Final csv preparation
+    smile_list = []
+    with connect(db_path) as db:
+        for a_row in db.select():
+            smile_list.append(a_row.smile)
+        if 'inchi' in a_row._keys:
+            gen_flag = 1
+        else:
+            gen_flag = 0
+    assert len(smile_list)>0, f'Input empty db or db has no SMILES'
 
     pyG_data_path = os.path.join(abs_data_path, 'input_pyG_data')
     db_2_pyG_data(db_path=db_path, properties=target_list, pyG_data_folder=pyG_data_path)
@@ -60,8 +86,13 @@ def pyG_infer_from_db(db_path, target_list, eval_ckpt_path, output_directory,
         a_model_path = os.path.join(eval_ckpt_path, f'{a_target}.pt')
 
         device = torch.device('cuda:0') if torch.cuda.is_available() else torch.device("cpu")
+        if torch.cuda.is_available():
+            device = torch.device('cuda:0')
+            ckpt = torch.load(a_model_path)
+        else:
+            device = torch.device("cpu")
+            ckpt = torch.load(a_model_path, map_location=torch.device('cpu'))
         trainer = pyG_trainer()
-        ckpt = torch.load(a_model_path)
         model.load_state_dict(ckpt['model_state_dict'])
         model.to(device=device)
 
@@ -71,11 +102,6 @@ def pyG_infer_from_db(db_path, target_list, eval_ckpt_path, output_directory,
         evaluation = pyG_inference_without_label(dump_info_path=output_directory, property=a_target)
         info = trainer.val(model=model, data_loader=DataLoader(test_dataset, infer_batch_size, shuffle=False),
                            energy_and_force=False, p=0, evaluation=evaluation, device=device)
-    # Final csv prepare
-    smile_list = []
-    with connect(db_path) as db:
-        for a_row in db.select():
-            smile_list.append(a_row.smile)
     cwd_2 = os.getcwd()
     os.chdir(abs_data_path)
     with open('input_smile.txt', 'w') as f:
@@ -83,21 +109,37 @@ def pyG_infer_from_db(db_path, target_list, eval_ckpt_path, output_directory,
             f.write(a_smile)
             f.write('\n')
     os.chdir(output_directory)
-    # print(os.getcwd())
     info_dict = {'Smiles': smile_list}
+    real_info_map = {}
     for a_target in target_list:
         a_info = np.load(f'{a_target}.npy')
-        # info_dict.update({a_target: a_info})
         info_dict.update({decorated_property_names[a_target]: a_info})
+        real_info_map.update({a_target: a_info})
     info_df = pd.DataFrame(info_dict)
-    info_df.to_csv('output_properties.csv')
+    info_df.to_csv('output_properties.csv', index=False)
     os.chdir(cwd_2)
+    with connect(db_path) as db, connect('temp.db') as db_w:
+        for idx, a_row in enumerate(db.select()):
+            an_atoms = a_row.toatoms()
+            data = {}
+            for a_target in target_list:
+                data.update({a_target:[float(real_info_map[a_target][idx])]})
+            if gen_flag:
+                # data.update({'ra_score': [a_row.data['ra_score'][0]]})
+                db_w.write(an_atoms, smile=a_row.smile, data=data, inchi=a_row.inchi)
+            else:
+                db_w.write(an_atoms, smile=a_row.smile, data=data)
+    os.remove(db_path)
+    os.rename(src='temp.db', dst=db_path)
+    edit_ase_db(db_path=db_path, properties=target_list)
+
+
 
 
 def pyG_infer(input_file_path, target, output_directory, eval_ckpt_path):
     target_list = []
     for a_target in target:
-        target_list.append(decorated_property_names_reverse[a_target.value])
+        target_list.append(decorated_property_names_reverse[a_target])
     print(target_list)
 
     abs_data_path = os.path.abspath(os.path.join(output_directory, 'data'))
@@ -110,10 +152,31 @@ def pyG_infer(input_file_path, target, output_directory, eval_ckpt_path):
     else:
         fail_smile_path = os.path.join(abs_data_path, 'fail_smile')
         smile_2_db(smile_path=input_file_path, db_path=db_path,
-                   properties=target_list, fail_smile_path=fail_smile_path)
+                   properties=target_list, fail_smile_path=fail_smile_path, maxAttempts=100)
     pyG_infer_from_db(db_path=db_path, target_list=target_list, eval_ckpt_path=eval_ckpt_path,
                       output_directory=output_directory, abs_data_path=abs_data_path)
 
+
+def infer_one_smile(target_smile: str, eval_ckpt_path: str):
+    cwd_ = os.getcwd()
+    os.makedirs('leftnet_infer', exist_ok=True)
+    os.chdir('leftnet_infer')
+    with open('input_smile', 'w') as f_w:
+        f_w.write(target_smile)
+    smile_2_db(smile_path='input_smile',
+               db_path=r'input.db',
+               fail_smile_path='failed_smile',
+               properties=list(decorated_property_names.keys()))
+    pyG_infer_from_db(db_path='input.db',
+                      target_list=list(decorated_property_names.keys()),
+                      eval_ckpt_path=eval_ckpt_path,
+                      output_directory=r'./')
+    info = pd.read_csv(r'output_properties.csv')
+    prop_arr = np.array([info['Binding energy(eV)'][0], info['Dielectric constant'][0], info['Viscosity (mPa*s)'][0],
+                         info['HOMO(eV)'][0], info['LUMO(eV)'][0]])
+    shutil.copy(src=r'output_properties.csv', dst=os.path.join(cwd_, r'input_smiles_properties.csv'))
+    os.chdir(cwd_)
+    return prop_arr
 
 
 
@@ -139,7 +202,7 @@ def pyg_infer_with_dpdispatcher(machine_info, cmdline):
                 "disk_size": 200,
                 "scass_type": machine_info['hardware'],
                 "platform": machine_info['platform'],
-                "image_name": "registry.dp.tech/dptech/prod-11729/uni-electrolyte-app:uni-electrolyte-app-bootstrap"
+                "image_name": "registry.dp.tech/dptech/prod-11729/gen_score_screen:1229"
             }
         }
     }