+++ src/uni_electrolyte/evaluator/inference/pyG_inference.py
@@ -28,9 +28,9 @@ class pyG_inference_test:
         y_pred_flatten = torch.flatten(y_pred).detach().cpu().numpy()
         y_true_flatten = torch.flatten(y_true).detach().cpu().numpy()
 
-        # if self.property in ['viscosity', 'dielectric_constant']:
-        #     y_true_flatten = pow(10, y_true_flatten)
-        #     y_pred_flatten = pow(10, y_pred_flatten)
+        if self.property in ['viscosity', 'dielectric_constant']:
+            # y_true_flatten = pow(10, y_true_flatten)
+            y_pred_flatten = pow(10, y_pred_flatten)
 
         with open("pred", "w") as pred:
             pred.writelines(list(map(lambda x: str(x) + "\n", y_pred_flatten)))